DownloadInvoices and ProcessTimesheets should be independent projects.
Break apart functionality into separate projects and share some common library code:
- configuration
- IMAP service basics (connect, disconnect)

+++ Config/ImapService.cs
@@ -11,10 +11,10 @@ namespace Config
             set { this["hostname"] = value; }
         }
 
-        [ConfigurationProperty("port", IsRequired = false, DefaultValue = (uint)993)]
-        public uint Port
+        [ConfigurationProperty("port", IsRequired = false, DefaultValue = 993)]
+        public int Port
         {
-            get { return (uint)this["port"]; }
+            get { return (int)this["port"]; }
             set { this["port"] = value; }
         }
     }
